Remove unused cached grpc api
