--- cmd/cached-client/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/gadget-inc/dateilager/pkg/cachedcli"
-
-func main() {
-	cachedcli.ClientExecute()
-}
