--- internal/pb/cache.proto
@@ -1,13 +1,0 @@
-syntax = "proto3";
-
-package pb;
-
-option go_package = "github.com/gadget-inc/dateilager/pkg/pb";
-
-service Cached {
-  rpc PopulateDiskCache(PopulateDiskCacheRequest) returns (PopulateDiskCacheResponse);
-}
-
-message PopulateDiskCacheRequest { string path = 1; }
-
-message PopulateDiskCacheResponse { int64 version = 1; };
