+++ pkg/client/context.go
@@ -4,8 +4,9 @@ import (
 	"context"
 )
 
-type clientCtxKey struct{}
-type cachedClientCtxKey struct{}
+type (
+	clientCtxKey struct{}
+)
 
 func FromContext(ctx context.Context) *Client {
 	client, ok := ctx.Value(clientCtxKey{}).(*Client)
@@ -17,16 +18,4 @@ func FromContext(ctx context.Context) *Client {
 
 func IntoContext(ctx context.Context, client *Client) context.Context {
 	return context.WithValue(ctx, clientCtxKey{}, client)
-}
-
-func CachedFromContext(ctx context.Context) *CachedClient {
-	client, ok := ctx.Value(cachedClientCtxKey{}).(*CachedClient)
-	if !ok {
-		return nil
-	}
-	return client
-}
-
-func CachedIntoContext(ctx context.Context, client *CachedClient) context.Context {
-	return context.WithValue(ctx, cachedClientCtxKey{}, client)
 }
